uploaded updated employee pages
This includes uploads folder and pages, previous work for adding a part,
editing a part and searching for a part. Employee login is also
included, however, that page does not work perfectly, so it would have
to be debugged. Inlcuded the img folder as well. Finally included the
whole updated party_shakers sql for the database.